Meeting_Notify_Agent/requirement.txt
uagents
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
